requirements.txt
Flask==2.3.3
pymongo==4.6.0
dnspython==2.4.2
selenium==4.15.2
pytest==7.4.3
Werkzeug==2.3.7
python-dotenv==1.0.0
